api/customer:  remove 'password' in res
