+++ app.js
@@ -36,12 +36,11 @@ function verify(req, res, next) {
 
 app.use("/", router);
 app.use("/api/customer", verify, require("./routes/customer.route"));
-// app.use("/api/customer", require("./routes/customer.route"));
+app.use("/api/employee", require("./routes/employee.route"));
+
 app.use("/api/account", verify, require("./routes/account.route"));
 
 app.use("/api/partnerbank", require("./routes/partnerbank.route"));
-// app.use('/client/mpbank', require("./routes/client.mpbank.route"));
-// app.use('/client/s2pbank', require("./routes/client.s2qbank.route"));
 
 app.use((req, res, next) => {
   res.status(404).send("NOT FOUND");
